requirements.txt
numpy==1.23.3
Pillow==10.0.0
pycocotools==2.0.6
torch==1.12.1
torchvision==0.15.2
